Added BG color and played with Div elements
I added some gradient background styling and played with the class selectors to create boxed with round corner around the recipe text.
Also added the return and homepage links as buttons to be made functional later with Javascript.
Added a ton of break tags to index.html to provide comfortable viewing of the gradient. Planning on adding a next button link in the header of all the recipe pages. changed header class styles.